==> notebooks/5.0 DL_Text.ipynb <==
## import et preparation des 3 df de textes

from joblib import dump,load

df = load("df_translated.joblib")
df.description = df.description.astype("str")
df.description_fr = df.description_fr.astype("str")
df.description_en = df.description_en.astype("str")
# fusion de désignation et description
df["texte_non_traduit"] = df.apply(lambda row: row["designation"] + " " + row["description"] if ((row["description"]!='nan') and (row["designation"] != row["description"])) else row["designation"], axis=1)

df["texte_traduit_fr"] = df.apply(lambda row: row["designation_fr"] + " " + row["description_fr"] if ((row["description_fr"]!='nan') and (row["designation_fr"] != row["description_fr"])) else row["designation_fr"], axis=1)

df["texte_traduit_en"] = df.apply(lambda row: row["designation_en"] + " " + row["description_en"] if ((row["description_en"]!='nan') and (row["designation_en"] != row["description_en"])) else row["designation_en"], axis=1)

#retrait des colonnes initiales description et designation
to_drop=['description','designation','productid', 'imageid','langues détectées', 'designation_fr', 'designation_en',
       'description_fr', 'description_en']
df = df.drop(columns = to_drop)
df_texte_non_traduit = df.drop(columns =['texte_traduit_fr', 'texte_traduit_en'])
df_texte_traduit_fr = df.drop(columns =['texte_non_traduit', 'texte_traduit_en'])
df_texte_traduit_en = df.drop(columns =['texte_non_traduit', 'texte_traduit_fr'])

dump(df_texte_non_traduit,"df_texte_non_traduit.joblib")

dump(df_texte_traduit_fr,"df_texte_traduit_fr.joblib")

dump(df_texte_traduit_en,"df_texte_traduit_en.joblib")

df_texte_non_traduit = load("df_texte_non_traduit.joblib")
df_texte_traduit_fr = load("df_texte_traduit_fr.joblib")
df_texte_traduit_en = load("df_texte_traduit_en.joblib")

## Prepa des données FR avant DL

df_texte_traduit_fr

import re
import unicodedata
from nltk.corpus import stopwords
from nltk.tokenize import word_tokenize


stop_words = stopwords.words('french')

# Converts the unicode file to ascii
def unicode_to_ascii(s):
    return ''.join(c for c in unicodedata.normalize('NFD', s)
        if unicodedata.category(c) != 'Mn')

def preprocess_sentence(w):
    w = unicode_to_ascii(w.lower().strip())
    # creating a space between a word and the punctuation following it
    # eg: "he is a boy." => "he is a boy ."
    w = re.sub(r"([?.!,¿])", r" \1 ", w)
    w = re.sub(r'[" "]+', " ", w)
    # replacing everything with space except (a-z, A-Z, ".", "?", "!", ",")
    w = re.sub(r"[^a-zA-Z?.!]+", " ", w)
    w = re.sub(r'\b\w{0,2}\b', '', w)

    # remove stopword
    mots = word_tokenize(w.strip())
    mots = [mot for mot in mots if mot not in stop_words]
    return ' '.join(mots).strip()

df_texte_traduit_fr.texte_traduit_fr = df_texte_traduit_fr.texte_traduit_fr.apply(lambda x :preprocess_sentence(x))
df_texte_traduit_fr.head()


from sklearn.preprocessing import LabelEncoder
le = LabelEncoder()
le.fit([  10, 2280,   50, 1280, 2705, 2522, 2582, 1560, 1281, 1920, 2403,
       1140, 2583, 1180, 1300, 2462, 1160, 2060,   40,   60, 1320, 1302,
       2220, 2905, 2585, 1940, 1301])

from sklearn.model_selection import train_test_split

features_fr = df_texte_traduit_fr.texte_traduit_fr


target_fr = df_texte_traduit_fr.prdtypecode
target_fr_encoded = le.transform(target_fr)
X_text_train, X_text_test, y_train, y_test = train_test_split(features_fr, target_fr_encoded, test_size=0.2, random_state=123)

import tensorflow as tf
# Définition du tokenizer
tokenizer = tf.keras.preprocessing.text.Tokenizer(num_words=10000)
# Mettre à jour le dictionnaire du tokenizer
tokenizer.fit_on_texts(X_text_train)


# Définition des dictionnaires
word2idx = tokenizer.word_index
idx2word = tokenizer.index_word
vocab_size = tokenizer.num_words

X_train = tokenizer.texts_to_sequences(X_text_train)

X_test = tokenizer.texts_to_sequences(X_text_test)

maxlen = 500
X_train = tf.keras.preprocessing.sequence.pad_sequences(X_train, maxlen=maxlen, padding='post', truncating='post')
X_test = tf.keras.preprocessing.sequence.pad_sequences(X_test, maxlen=maxlen, padding='post', truncating='post')


from tensorflow.keras import Sequential
from tensorflow.keras.layers import Embedding, Dense, GlobalAveragePooling1D, RNN, GRUCell, Dropout
embedding_dim = 200

model = Sequential()
model.add(Embedding(10000, embedding_dim))
model.add(RNN(GRUCell(128), return_sequences=True))
model.add(Dropout(0.3))
model.add(GlobalAveragePooling1D())
model.add(Dense(256, activation='relu'))
model.add(Dropout(0.3))
model.add(Dense(27, activation='softmax'))

model.summary()

model.compile(optimizer='adam',
              loss='sparse_categorical_crossentropy',
              metrics=['accuracy'])


history = model.fit(X_train, y_train,
    batch_size = 64,
    epochs=10,
    validation_data = [X_test, y_test])

model.evaluate(X_test, y_test, batch_size=128)

model.save("DL_text")
